requirements.txt
Django>=5.2.1
python-dotenv
imap_tools